helpers tests
